+++ tubular/_registry.py
@@ -1,12 +1,22 @@
+from tubular.base import BaseTransformer
+
 CLASS_REGISTRY = {}
 
 
-def register(cls: type) -> type:
-    """Add class object to registry(dict).
+def register(cls: BaseTransformer) -> BaseTransformer:
+    """Add transformer to registry dict.
 
-    Returns
+    Returns:
     -------
-    cls - class object.
+    cls - transformer
+
+    Example:
+    -------
+    >>> @register
+    ... class MyTransformer(BaseTransformer):
+    ...     pass
+    >>> CLASS_REGISTRY["MyTransformer"]
+    <class 'MyTransformer'>
 
     """
     CLASS_REGISTRY[cls.__name__] = cls