+++ tubular/mapping.py
@@ -27,6 +27,7 @@ if TYPE_CHECKING:
     from narwhals.typing import IntoDType
 
 
+@register
 class BaseMappingTransformer(BaseTransformer):
     """Base Transformer Extension for mapping transformers.
 
@@ -248,6 +249,7 @@ class BaseMappingTransformer(BaseTransformer):
         return _return_narwhals_or_native_dataframe(X, return_native)
 
 
+@register
 class BaseMappingTransformMixin(BaseTransformer):
     """Mixin class to apply mappings to columns method.